Refactor Menu scrolling into MenuDisplay
This commit moves the scrolling handling for Menus into the MenuDisplay
class. This is necessary because the menu itself does not have the
layout information necessary to scroll or render the current page of
items, only the renderer (MenuDisplay) does.

Below, "Tile menu" means a menu using a MenuDisplayTile instance;
likewise "Text menu" means a menu using a MenuDisplayText instance.

There are a number of other changes+improvements this commit makes:

 - the two MenuDisplay derived classes have been moved into menu.cc
 - an MF_USE_TWO_COLUMNS flag has been added for TILES_LOCAL builds
 - Tile menus force a page break before a duplicate hotkey
 - Text menus limit their pagesize to 52 rows if duplicate hotkeys are
   detected. This approach keeps the Text menu scrolling/layout code simple.
 - showing the --more-- line now works correctly with menus with
   multiple parts delimited by MEL_END_OF_SECTION
 - scrolling formatted_scrollers with home/end keys now stays within the
   current section. This breaks the ? help menu, but I've added a small
   hack for that. It's still broken on webtiles, but it was like that
   alrady.

Removal of functionality:

 - Tiles menus can no longer be scrolled by line, as the notion of
   scrolling by line does not work well when combined with two-column
   layouts. Scrolling to an item scrolls to the page containing that item.