Remove unneeded tiles_crt_control instances
All of these instances are followed by code that uses
show_description(), which already pushes and pops a menu on webtiles
with code internal to Menu. The second string parameter is used to
add a class to the menu HTML element to permit per-menu customization,
but none of these were in use. (class: "menu_" + tag)